pyproject.toml
[tool.poetry]
name = "review-app-scan"
version = "1.3.0"
description = ""
authors = ["Matthew Cane <matthew.cane@citizensadvice.org.uk>"]
readme = "README.md"
packages = [{ include = "review_app_scan" }]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
tqdm = "^4.66.3"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scan = "review_app_scan.__main__:main"
